pyproject.toml
[tool.poetry]
name = "shipay-backend-challenge"
version = "0.1.0"
description = "Projeto implementado para o desafio de Desenvolvedor Júnior da Shipay"
authors = ["Lnvictor <lnvictor@github.com>"]
license = "GNU"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.29"
psycopg2-binary = "^2.9.3"
Flask = "^2.0.2"
bcrypt = "^3.2.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
ipdb = "^0.13.9"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
